requirements.txt
isort==4.2.5
pluggy==0.4.0
py==1.4.32
pytest==3.0.5
PyYAML==3.12
six==1.10.0
tox==2.5.0
virtualenv==15.1.0
